requirements.txt
pandas
numpy
Flask
pymongo
flask-pymongo
python-dotenv

-e .